change navigation structure for setting screen
